Refactor: Pass env into completion in remote control
In preparation for suggesting modules to import, it became apparent
we'd need mroe than just the document string going into the completion
call. The completion environment seems like a good choice, as it
offers the document, analysis, and a host of other features related to
figuring out which context you're in.

This also allowed us to push elixir_sense specific logic into the
completions module in remote control, further locating it closer to
its call site.
